Добавлна папка игры, перенесены все игровые файлы в папку
