+++ README.md
@@ -1,3 +1,10 @@
 # Project-RPG-
 Проект для лабораторной работы 12
 Тема проекта "РПГ игра"
+
+Команда 
+Дима Трубач
+Никита Кутейко
+Ломако Александр
+Песецкий Илья
+Песецкий Никита
